Model API and Avatar api fixed

+++ KoGaMaTools.Native/kogama-tools/UI/Tabs/BuildUITab.cpp
@@ -4,7 +4,7 @@
 void KoGaMaTools::UI::BuildUITab::Setup(KoGaMaTools::UI::MainUI& ui)
 {
 	ui.AddComponentType("Build");
-	auto i = ui.components.size() - 1;
+	auto i = static_cast<int>(ui.components.size() - 1);
 
 	namespace S = KoGaMaTools::Services;
