+++ KoGaMaTools.Native/kogama-tools/UI/Tabs/ConfigUITab.cpp
@@ -4,7 +4,7 @@
 void KoGaMaTools::UI::ConfigUITab::Setup(KoGaMaTools::UI::MainUI& ui)
 {
 	ui.AddComponentType("Config");
-	auto i = ui.components.size() - 1;
+	auto i = static_cast<int>(ui.components.size() - 1);
 
 	namespace S = KoGaMaTools::Services;
