+++ KoGaMaTools.Native/kogama-tools/UI/Tabs/PvPUITab.cpp
@@ -4,7 +4,7 @@
 void KoGaMaTools::UI::PvPUITab::Setup(KoGaMaTools::UI::MainUI& ui)
 {
 	ui.AddComponentType("PvP");
-	auto i = ui.components.size() - 1;
+	auto i = static_cast<int>(ui.components.size() - 1);
 
 	namespace S = KoGaMaTools::Services;
